Set MDNS as default registry
